KE-41667 [FOLLOW UP] fix error (#614)
