+++ sql/hive-thriftserver/src/main/scala/org/apache/spark/sql/hive/thriftserver/SparkSQLCLIDriver.scala
@@ -99,11 +99,11 @@ private[hive] object SparkSQLCLIDriver extends Logging {
 
     sessionState.in = System.in
     try {
-      HiveUtils.shimSessionState(sessionState,
+      shimSessionState(sessionState,
         "out", new PrintStream(System.out, true, UTF_8.name()))
-      HiveUtils.shimSessionState(sessionState,
+      shimSessionState(sessionState,
         "info", new PrintStream(System.err, true, UTF_8.name()))
-      HiveUtils.shimSessionState(sessionState,
+      shimSessionState(sessionState,
         "err", new PrintStream(System.err, true, UTF_8.name()))
     } catch {
       case e: UnsupportedEncodingException => System.exit(3)
@@ -183,11 +183,11 @@ private[hive] object SparkSQLCLIDriver extends Logging {
     // will set the output into an invalid buffer.
     sessionState.in = System.in
     try {
-      HiveUtils.shimSessionState(sessionState,
+      shimSessionState(sessionState,
         "out", new PrintStream(System.out, true, UTF_8.name()))
-      HiveUtils.shimSessionState(sessionState,
+      shimSessionState(sessionState,
         "info", new PrintStream(System.err, true, UTF_8.name()))
-      HiveUtils.shimSessionState(sessionState,
+      shimSessionState(sessionState,
         "err", new PrintStream(System.err, true, UTF_8.name()))
     } catch {
       case e: UnsupportedEncodingException => System.exit(3)
@@ -306,6 +306,15 @@ private[hive] object SparkSQLCLIDriver extends Logging {
     System.exit(ret)
   }
 
+  def shimSessionState(state: SessionState, //
+                       field: String, stream: java.io.OutputStream): Unit = {
+    val fieldObj = state.getClass.getField(field)
+    val fieldValue = fieldObj.getType
+      // eg. PrintStream -> SessionStream (CDP 7.1 specified)
+      .getConstructor(classOf[java.io.OutputStream])
+      .newInstance(stream)
+    fieldObj.set(state, fieldValue)
+  }
 
   def isRemoteMode(state: CliSessionState): Boolean = {
     //    sessionState.isRemoteMode
